feat: Implement UI
First steps of implementing a decent-looking UI for our add-on using React Spectrum.

Note that this plans to use the currently in Beta `Toast` components of React Spectrum, which might change in the future. Docs for this can be found on <https://react-spectrum.adobe.com/react-spectrum/Toast.html>.

Note that we used a somewhat hacky `closeFunction` module level variable to ensure that only one large toast gets shown at a time. Not the prettiest, but there are no better APIs for this at the moment.